added file upload\remove functions
